+++ R/helper.R
@@ -5,23 +5,38 @@
 #' @export
 #' 
 #' @param shiny.tag A shiny input or output object
+#' @param filename A character string of the name of the helpfile you want to display
+#' @param size Either 's', 'm' or 'l' - the size of the modal dialog to display
 #' @param icon An icon created with \code{shiny::icon()}
 #' @param class A custom CSS class - defaults to shiny-helper-question
-#' @param colour Any valid CSS colour for the icon (if you pass a colour 
-#'   as a \code{style} argument, \code{colour} will be overruled)
+#' @param icon_colour Any valid CSS colour for the icon (if you pass a colour 
+#'   as a \code{style} argument, \code{icon_colour} will be overruled)
 #' @param ... Other arguments to pass to shiny::actionButton
 #' 
 #' @examples 
 #' helper(shiny::actionButton(inputId = "button1", label = "Click me!"))
-#' helper(shiny::plotOutput(outputId = "plot1"))
-helper <- function(shiny.tag, icon = shiny::icon("question-circle-o"),
+#' helper(shiny::plotOutput(outputId = "plot1"), size = "l", filename = "PlotHelp", 
+#'        icon_colour = "green")
+helper <- function(shiny.tag, 
+                   filename = NULL,
+                   size = "m",
+                   icon = shiny::icon("question-circle-o"),
                    class = "shiny-helper-question", 
-                   colour = NULL,
+                   icon_colour = NULL,
                    ...){
+  
+  if (!(size %in% c("s", "m", "l")))
+    stop("size must be in c('s', 'm', 'l')")
   
   id <- get_id(shiny.tag)
   
-  if (!is.null(colour)) colour <- paste0("color: ", colour, ";")
+  if (is.null(filename)) filename <- id
+  if (grepl("\\.md$", filename)) {
+    filename <- sub("\\.md$", "", filename)
+    message("No need to include '.md' in filenames")
+  }
+  
+  if (!is.null(icon_colour)) icon_colour <- paste0("color: ", icon_colour, ";")
   
   shiny::tagList(
     shiny::singleton(
@@ -29,11 +44,12 @@ helper <- function(shiny.tag, icon = shiny::icon("question-circle-o"),
     ),
     shiny::div(class = "shiny-helper-container",
                shiny.tag, 
-               shiny::actionButton(inputId = paste0(id, "-shinyhelper-msg"), 
+               shiny::actionButton(inputId = paste(id, filename, size, 
+                                                   "shinyhelper", sep = "---"), 
                                    label = NULL,
                                    icon = icon,
                                    class = class,
-                                   style = colour,
+                                   style = icon_colour,
                                    ...)
     )
   )