+++ inst/example/server.R
@@ -8,9 +8,7 @@ server <- function(input, output, session) {
   
   # use helpfiles
   observe_helpers(input = input, output = output,
-                  help_dir = "helpfiles", 
-                  sizes = list(kmeans = "l", ycol = "s"),
-                  default_size = "m")
+                  help_dir = "helpfiles")
   
   # demo app for kmeans clustering taken from 
   # https://shiny.rstudio.com/gallery/kmeans-example.html